Add 'github_deep_clone' instruction
This PR implements the deep clone tip in [Cirrus Docs], using
starlak.

It should be a fairly common use case, so maybe it is good to have it
in the helpers repository.

Some build tools have high expectations about the state of the `.git`
directory (e.g., when automatically deriving version tags from branch
and tag names), that might not play well with `go-git` (probably because
of the single-branch behaviour?).

[Cirrus Docs] show a very useful workaround for this, that consists in
defining a custom `clone_script`. The drawback of this approach is that
different `$SHELL`s (and therefore OS, i.e.: WINDOWS), don't support the
same syntax for conditionals.

Fortunatelly, we can use Starlak to run the conditionals and bypass the
need for shell scripting them.

[Cirrus Docs]: https://cirrus-ci.org/guide/tips-and-tricks